fix: allow setting bnode_depth to zero
allow setting blank node depth to zero in generate_item_construct
previously this was not respected. and a bnode_depth of zero would still
result in retrieving blank nodes to a depth of one.

The issue being that for search queries, this meant the entire graph
was being traversed every time, no matter what, seriously impacting
performance.