ci: migrate from python-semantic-release to semantic-release
Python semantic release was only used to auto update the version
variable in pyproject.toml, but this does not work with branch
protection rules and so is not neccessary anymore. Thus it is better to
use the more common semantic-release action which works with branch
protection rules as it does not need to commit any files as part of the
release.