feat: allow merging of prez config from a local dir, mainly to suppor… (#432)
* feat: allow merging of prez config from a local dir, mainly to support mounting in custom config in the docker container

* add missing entrypoint

* fix: entrypoint args

* save

* remove bash from dockerfile; sh is more appropriate for alpine
changse to entrypoint/dockerfile to support mounting in volumes read only such that the mounted filesystem is not modified to include default prez config

* black

* include rdf xml as an input type for queryables to get around rdflib parsing recursion limits in long rdf lists

* feat: wrap cql datetime and bounding box expressions in filter exists (#410)

* fix: implement <> CQL operator

* simplify and improve error handling for CQL GET

* update tests to account for new filter exists pattern, add further tests for combinations of datetime/spatial/cql

* fix: include pagination query param validation and remove duplicated tssp for spatial filters

* add plan

* wip

* update spatial filter

* revert a lot of changes:
FILTER EXISTS is now only scoped to the GraphDB formatted geospatial queries.

While there is potentially value in FILTER EXISTS for BGP that do not have ?focus_node as their subject, these triples are not generated in the CQL/datetime/bbox filters, other than geo triples. The geo triples did not generally show a performance benefit in testing for the geosparql query format. As such FILTER EXISTS is only introduced for the GraphDB spatial query format. The SHACL path expressions used for CQL queryables all have ?focus_node as the subject and use the path syntax. With this understanding it is now not necessary to introduce FILTER EXISTS generally *in relation to CQL/datetime/spatial* filters. It could still be beneficial to introduce FILTER EXISTS for BGPs that are from endpoints.

* remove now inaccurate plan

* black

* perf: add filter exists wrapping to `x rdf:type y` triples generated in subselects. This improves the performance of listing queries and link generation. (#433)

* feat: add qualified attribution / hadRole CQL function (#423)

* feat: add qualified attribution / hadRole CQL function

* black code

---------

Co-authored-by: lalewis1 <lawson@kurrawong.ai>